requirements.txt
tensorflow
fastapi
uvicorn
uvicorn[standard]
gunicorn
python-multipart
pillow
matplotlib
numpy
opencv-python